Rename folder to blocks
