+++ server/app.js
@@ -13,7 +13,7 @@ const errorHandlers = require('./errorHandlers');
 
 // routers
 const clientRouter = require('./components/client/clientRouter');
-const blocksRouter = require('./components/block/blocksRoutes');
+const blocksRouter = require('./components/blocks/blocksRoutes');
 
 const app = express();
